refactor: extract LLM run core
